Moved the searchbar files into the Searchbar folder
